# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the Developer Portal, providing endpoints for managing organizations, teams, projects, components, and deployments.",
        "title": "Developer Portal Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/alert-storage/alerts/{project}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve alerts for a specific project with optional filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert-storage"
                ],
                "summary": "Get alerts by project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name (must be one of cis2, usrv, cloud_automation)",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (min 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page (min 1, max 100)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by severity (e.g., critical, warning, info)",
                        "name": "severity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region label",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by landscape",
                        "name": "landscape",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "firing",
                            "resolved"
                        ],
                        "type": "string",
                        "description": "Filter by status (firing or resolved)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component label",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by alert name",
                        "name": "alertname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter alerts after this time (RFC3339 format)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter alerts before this time (RFC3339 format)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved alerts",
                        "schema": {
                            "$ref": "#/definitions/client.AlertStoragePaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alert-storage/alerts/{project}/filters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve available filter values for alerts in a specific project. Returns a dynamic map where keys are filter names (e.g., alertname, severity, status, landscape, region) and values are arrays of available options. Use query parameters to narrow down results.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert-storage"
                ],
                "summary": "Get available filter values for alerts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name (must be one of cis2, usrv, cloud_automation)",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by severity level (e.g., critical, warning, info)",
                        "name": "severity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by landscape (e.g., production, staging, development)",
                        "name": "landscape",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "firing",
                            "resolved"
                        ],
                        "type": "string",
                        "description": "Filter by alert status (firing or resolved)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by alert name",
                        "name": "alertname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region (e.g., us-east-1, eu-west-1)",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by component label",
                        "name": "component",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter alerts starting from this time (RFC3339 format)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter alerts up to this time (RFC3339 format)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved filter values as a dynamic map",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alert-storage/alerts/{project}/{fingerprint}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific alert by its unique fingerprint within a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert-storage"
                ],
                "summary": "Get alert by fingerprint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name (alphanumeric, dashes, underscores)",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Alert fingerprint (hexadecimal string, max 128 chars)",
                        "name": "fingerprint",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved alert",
                        "schema": {
                            "$ref": "#/definitions/client.AlertStorageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alert-storage/alerts/{project}/{fingerprint}/label": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update or add a label for a specific alert",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert-storage"
                ],
                "summary": "Update alert label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project name (alphanumeric, dashes, underscores)",
                        "name": "project",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Alert fingerprint (hexadecimal string, max 128 chars)",
                        "name": "fingerprint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Label update request with 'key' and 'value' fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Label updated successfully",
                        "schema": {
                            "$ref": "#/definitions/client.UpdateLabelResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters or body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Alert not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alert-storage/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all available projects configured in the alert storage system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alert-storage"
                ],
                "summary": "Get all alert storage projects",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved projects",
                        "schema": {
                            "$ref": "#/definitions/client.ProjectsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "client.AlertStoragePaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/client.AlertStorageResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "client.AlertStorageResponse": {
            "type": "object",
            "properties": {
                "alertname": {
                    "type": "string"
                },
                "annotations": {
                    "type": "object",
                    "additionalProperties": true
                },
                "createdAt": {
                    "type": "string"
                },
                "endsAt": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": true
                },
                "landscape": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "startsAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "client.ProjectsResponse": {
            "type": "object",
            "properties": {
                "projects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "client.UpdateLabelResponse": {
            "type": "object",
            "properties": {
                "fingerprint": {
                    "type": "string"
                },
                "label": {
                    "type": "object",
                    "properties": {
                        "key": {
                            "type": "string"
                        },
                        "value": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}